# goios-peer/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/wda/session": {
            "post": {
                "description": "Create a new WebDriverAgent session for the specified device",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "WebDriverAgent"
                ],
                "summary": "Create a new WDA session",
                "parameters": [
                    {
                        "description": "WebDriverAgent Configuration",
                        "name": "config",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.WdaConfig"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.WdaSession"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.GenericResponse"
                        }
                    }
                }
            }
        },
        "/wda/session/{sessionId}": {
            "get": {
                "description": "Get a WebDriverAgent session by sessionId",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "WebDriverAgent"
                ],
                "summary": "Get a WebDriverAgent session",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session ID",
                        "name": "sessionId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.WdaSession"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.GenericResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a WebDriverAgent session by sessionId",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "WebDriverAgent"
                ],
                "summary": "Delete a WebDriverAgent session",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session ID",
                        "name": "sessionId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.WdaSession"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.GenericResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.GenericResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "api.WdaConfig": {
            "type": "object",
            "required": [
                "bundleId",
                "testBundleId",
                "xcTestConfig"
            ],
            "properties": {
                "args": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "bundleId": {
                    "type": "string"
                },
                "env": {
                    "type": "object",
                    "additionalProperties": true
                },
                "testBundleId": {
                    "type": "string"
                },
                "xcTestConfig": {
                    "type": "string"
                }
            }
        },
        "api.WdaSession": {
            "type": "object",
            "required": [
                "config",
                "sessionId",
                "udid"
            ],
            "properties": {
                "config": {
                    "$ref": "#/definitions/api.WdaConfig"
                },
                "sessionId": {
                    "type": "string"
                },
                "udid": {
                    "type": "string"
                }
            }
        },
...
    }
}